Fix importing Nexus into IPython.
Importing Nexus into IPython stops with an error:
"NameError: name 'exit' is not defined"
Nexus is attempting to store the exit routine for later use.
In IPython, 'exit' is of type IPython.core.autocall.ExitAutoCall, which
is apparently not defined for imported classes.
In the regular Python interpreter, 'exit' if of type 'site.Quitter'.

Using 'sys.exit' (which is a built-in function) instead keeps IPython
happy, and should have the same effect with regular Python.