Only import matplotlib if qmca is plotting.
Addresses #178.
Move the matplotlib initialization to a function.  But then all the imported names
are in function's local namespace.  Fix that by declaring the variables global.

Rearrange the logic so the variables are defined in both branches of the try/except.
This keeps pyflakes from complaining about redefined variables.

Add calls to the 'import_matplotlib' function in the code paths where plotting will occur.